refactor(swing-store): refactor swingstore into better pieces
This creates new files to house `importSwingStore`,
`makeSwingStoreExporter`, and the kvStore. All were previously in
`swingStore.js`, which is now a shell that gathers together the other
components. The module exports have been consolidated into
`src/index.js`, and types are now exported by a
`types.d.ts`. No behavioral changes.
