+++ packages/swing-store/src/types.d.ts
@@ -1,0 +1,14 @@
+export type {
+  SwingStore,
+  SwingStoreKernelStorage,
+  SwingStoreHostStorage,
+} from './src/swingStore.js';
+export type { KVStore } from './src/kvStore.js';
+export type { BundleStore } from './src/bundleStore.js';
+export type {
+  SnapStore,
+  SnapshotResult,
+  SnapshotInfo,
+} from './src/snapStore.js';
+export type { TranscriptStore } from './src/transcriptStore.js';
+export type { SwingStoreExporter, ExportMode } from './src/exporter.js';
